proyecto_feria_django/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.35.41
botocore==1.35.41
CacheControl==0.14.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.1
dj-database-url==2.2.0
Django==5.1.2
django-cors-headers==4.5.0
django-storages==1.14.4
djangorestframework==3.15.2
firebase-admin==6.5.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
kombu==5.4.2
msgpack==1.1.0
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pdf2image==1.17.0
pillow==11.0.0
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
